Make our pyparted usage more robust (Kingson DataTraveller workaround).
For some reason calling disk.commit() to mark a partition as bootable fails on
Kingson DataTravellers with a "Cannot commit to disk" DiskException.  However,
the partition is successfully marked as bootable after this error, so we should
make sure this error isn't fatal, but make the exception available for further
debugging.