Fix the filename parsing when retrieving available images from dl.fp.o
Cutting off 6 characters from the end instead of 4 drops the rather useless
trailing -1 (it's an artifact of Koji-based building of lives, AIUI).

The chunks.pop / chunks.insert dodge takes the last field from the filename
after trimming and splitting (which is the release version, and hopefully will
stay that way for a bit...) and puts it into the list at position 1. So this
is only dependent on the *last* field always being the release, and the
*first* field always being the distribution name - other fields can jiggle
around and this will still DTRT.